test plate_db with database inheritance, ok
