+++ BioPlate/database/plate_db.py
@@ -1,12 +1,10 @@
-import BioPlate.database.database_function_import as dfi
+from BioPlate.database.database import Database
 from sqlalchemy import Column, Integer, String, Float, exc
 
 
-class PlateDB:
+class PlateDB(Database):
 
-    database_name = 'plate.db'
-
-    class PlateDatabase(dfi.Base):
+    class PlateDatabase(Database.Base):
         """
         Database for plate
         """
@@ -36,17 +34,9 @@ class PlateDB:
             else:
                 return f"<plate N°{self.id} : {self.numWell}-{self.numColumns}-{self.numRows}>"
 
-    @classmethod
-    def from_database_name(cls, dbname):
-        cls.database_name = dbname
-        plate = cls(db_name=dbname)
-        return plate
 
-    def __init__(self, db_name = database_name):
-        self.db_name = db_name
-        self.sqlalchemypath, self.engine, self.session = dfi.eng_sess(self.db_name)
-        self.table_create = dfi.create_table(self.sqlalchemypath)
-
+    def __init__(self, db_name = 'plate.db'):
+        super().__init__(self.PlateDatabase, db_name)
 
     def add_plate(self, numWell, numColumns, numRows, name=None, surfWell=None, maxVolWell=None, workVolWell=None,
                   refURL=None):
@@ -63,7 +53,7 @@ class PlateDB:
         :param refURL:
         :return: Nothing
         """
-        already_exist = self.session.query(self.PlateDatabase).filter_by(
+        already_exist = self.session.query(self.database_class).filter_by(
                 numWell = numWell,
                 numColumns = numColumns,
                 numRows=numRows,
@@ -74,7 +64,7 @@ class PlateDB:
                 refURL=refURL).count()
 
         if not already_exist:
-            new_entry = self.PlateDatabase(
+            new_entry = self.database_class(
                     numWell=numWell,
                     numColumns=numColumns,
                     numRows=numRows,
@@ -91,84 +81,19 @@ class PlateDB:
         else:
             return None
 
-    def get_plate(self, args, key='numWell'):
-        """
+    def update_plate(self, dict_update, args, key="numWell"):
+        return super().update(dict_update, args, key=key)
 
-        :param session: sqlalchemy session
-        :param args: args to search of
-        :param key: column name in the database
-        :return:
-        """
-        plt = self.session.query(self.PlateDatabase).filter(getattr(self.PlateDatabase, key) == args).all()
-        self.session.close()
-        return plt
-
-    @property
+    def delete_plate(self, args, key="numWell"):
+        return super().delete(args, key=key)
+       
+    def get_plate(self, args, key="numWell"):
+        return super().get(args, key=key)
+          
     def get_all_plate(self):
-        """
-        get list of plate in the database
-        :param session: Sqlalchemy session
-        :param numWell: number of well in a plate (INT)
-        :return: a list of plate object
-        """
-        all_plate = self.session.query(self.PlateDatabase).all()
-        self.session.close()
-        return all_plate
-
-    def delete_plate(self, args, key='numWell'):
-        """
-
-        :param session: sqlalchemy session
-        :param args: args to search of
-        :param key: column name in the database
-        :return:
-        """
-        try :
-            dplt = self.session.query(self.PlateDatabase).filter(getattr(self.PlateDatabase, key) == args).one()
-            numwell = dplt.numWell
-            self.session.delete(dplt)
-            self.session.commit()
-            self.session.close()
-            return f"plate with {args} {key} deleted"
-        except exc.SQLAlchemyError:
-            return "Use a more specific key to delete the plate"
+        return super().get_all()
 
 
-if __name__ == '__main__':
-    #PlateDB.from_database_name('plate2.db')
-    pdb = PlateDB()
-    #pdb = PlateDB.from_database_name('plate2.db')
-    pdb.add_plate(numWell=96,
-              numColumns=12,
-              numRows=8,
-              surfWell=0.29,
-              maxVolWell=200,
-              workVolWell=200,
-              refURL='https://csmedia2.corning.com/LifeSciences/Media/pdf/cc_surface_areas.pdf')
-    pdb.add_plate(numWell=6,
-              numColumns=3,
-              numRows=2,
-              surfWell=9.5,
-              maxVolWell=2000,
-              workVolWell=2000,
-              refURL='https://csmedia2.corning.com/LifeSciences/Media/pdf/cc_surface_areas.pdf')
-    pdb.add_plate(numWell=24,
-              numColumns=6,
-              numRows=4,
-              surfWell=0.33,
-              maxVolWell=400,
-              workVolWell=400,
-              refURL='https://csmedia2.corning.com/LifeSciences/Media/pdf/cc_surface_areas.pdf')
-    pdb.add_plate(numWell=96,
-              numColumns=10,
-              numRows=6,
-              surfWell=0.29,
-              maxVolWell=200,
-              workVolWell=200,
-              refURL='https://csmedia2.corning.com/LifeSciences/Media/pdf/cc_surface_areas.pdf')
-    print(pdb.get_plate(96))
-    p6 = pdb.get_plate(400, key='maxVolWell')
-    #print(pdb.delete_plate(10, key="numColumns"))
-    print(p6)
-    test = pdb.get_all_plate
-    print(test)
+
+
+
